Refactor: Api logic
